requirements.txt
Flask==1.1.3
Flask-Cors==3.0.10
flask-restplus==0.12.1
gevent==21.12.0
mariadb==1.0.10
MarkupSafe==2.0.1
pytest==7.1.1
pytest-cov==3.0.0
python-dotenv==0.20.0
requests==2.27.1
Werkzeug==0.16.1
